+++ src/BarinakSistemi2/Musteri.java
@@ -8,13 +8,13 @@ public class Musteri extends Person {
     private String address;
     private String phone;
 
-    public Musteri(int id, String name, String email, String password, String customerId, String address, String phone) {
+    public Musteri(long id, String name, String email, String password, String customerId, String address, String phone) {
         super(id, name, email, password);
         this.customerId = customerId;
         this.address = address;
         this.phone = phone;
     }
-
+    
     public void setCustomerId(String customerId) {
     	this.customerId = customerId;
     }
@@ -24,7 +24,7 @@ public class Musteri extends Person {
     public void setPhone(String phone) {
     	this.phone = phone;
     }
-
+    
     public String getCustomerId() {
     	return customerId;
     }
@@ -34,7 +34,7 @@ public class Musteri extends Person {
     public String getPhone() {
     	return phone;
     }
-
+    
     @Override
     public String getInfo() {
         return "Müşteri: " + getName() + " | Tel: " + phone;
@@ -51,6 +51,6 @@ public class Musteri extends Person {
 
     public List<Hayvan> filtrele(String tur, int yas) {
         System.out.println(tur + " türünde ve " + yas + " yaşındaki hayvanlar filtreleniyor...");
-        return new ArrayList<>();
+        return new ArrayList<>(); 
     }
 }
